requirements.txt
# Python Dependencies for EDA Agent Template

# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.3.0

# Guardrails for structured output validation
guardrails-ai>=0.5.0

# UI Framework
streamlit>=1.28.0
click>=8.1.0

# Data Processing
pydantic>=2.5.0
pyyaml>=6.0.1
jsonschema>=4.20.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0
requests>=2.31.0
openai>=1.0.0
httpx>=0.24.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0

